requirements.txt
SQLAlchemy==1.3.9
bleach==3.1.0
Flask_Login==0.5.0
Flask_SQLAlchemy==2.4.1
Flask==1.1.1
Werkzeug==0.16.0
Flask_JWT_Extended==3.24.1
Pillow==7.1.2
markdown
pymysql
gunicorn